# Refactor parsing to improve startup performance