--- a/snapshot/gui/compare.py
+++ b/snapshot/gui/compare.py
@@ -454,7 +454,8 @@ class SnapshotPvTableModel(QtCore.QAbstractTableModel):
             macros = file_data["meta_data"].get("macros", dict())
 
         pvs_list_full_names = dict()  # PVS data mapped to real pvs names (no macros)
-        for pv_name_raw, pv_data in file_data["pvs_list"].items():
+        pvs_list, _, _ = self.snapshot.parse_from_save_file(file_data['file_path'])
+        for pv_name_raw, pv_data in pvs_list.items():
             pvs_list_full_names[SnapshotPv.macros_substitution(pv_name_raw, macros)] = pv_data
 
         return pvs_list_full_names