As agreed on the meetings on 24. 01. and 28. 01. 2020, I picked from
the set of proposed changes one with significant impact on the
performance of the snapshot tool.

The goal is reducing the time required for analyzing snapshot files on
program startup. This is achieved by only extracting the metadata of
each snapshot. Loading actual data is defered for later when it is
required to display or restore a snapshot. This also reduces memory
usage as previously, all data was kept in memory, while resources are
now released when data is not viewed anymore.

A side effect of defered loading is that only metadata parsing errors
are caught at startup time while data parsing errors are caught when
the file is loaded. As before, PVs with invalid data are simply
ignored.

During testing, an already existing bug was found and fixed where
editing metadata did not work unless the `cwd` of the program was in
the snapshot directory.

The effect of the change was tested with the `SF_settings` set of
snapshots where most of the snapshots have dates between 20180203
and 20200124. All files and the snapshot tool itself resided on an NFS
share. Measured was total startup time from pressing Return until the
window was displayed. Three cases were considered:

  1. Cache warm, PVs disconnected. Total startup time was reduced from
     38 seconds to 3 seconds. The cache warm case also reflects the
     improvement in the behaviour of the `Refresh` button, which
     completed in about 1 second. Memory usage was decreased from
     4.4 GB to 117 MB. Memory usage increases with the number of
     snapshots selected for comparison, but decreases again when
     snapshots are deselected.

  2. Cache cold (`echo 3 > /proc/sys/vm/drop_caches`), PVs
     disconnected. Startup time was reduced from 65 seconds to 15
     seconds.

  3. Cache cold, 1000 PVs connected and updating at 100 Hz. Startup
     time was 85 seconds before the change and 33 seconds after. When
     running, the GUI was of course sluggish because of high monitor
     rate; this is not affected by this PR.