@simongregorebner 2020-02-05
This looks good, we can see some improvements in the startup time in the SwissFEL setup.
Will merge this in...